django/requirements.txt
django-enumfields==0.10.1
django-extensions==2.0.7
django-polymorphic==2.0.2
django-tagulous==0.13.2
django-webpack-loader==0.6.0
Django==1.11.13
djangorestframework==3.8.2
djoser==1.3.2
gitpython==2.1.10
graphene-django>=2.0
gunicorn==19.8.1
ipdb==0.10.1
ipython==5.1.0
plumbum==1.6.6
psycopg2==2.6.2
pudb==2018.1
pytest-django==3.1.2
python-memcached==1.59
pytz==2018.4
PyYAML==3.12
requests==2.18.4
